+++ gems/aws-sdk-core/lib/aws-sdk-core/json/json_engine.rb
@@ -2,16 +2,18 @@
 
 module Aws
   module Json
-    class JSONEngine
+    module JSONEngine
+      class << self
+        def load(json)
+          JSON.parse(json)
+        rescue JSON::ParserError => e
+          raise ParseError.new(e)
+        end
 
-      def self.load(json)
-        JSON.load(json)
+        def dump(value)
+          JSON.dump(value)
+        end
       end
-
-      def self.dump(value)
-        JSON.dump(value)
-      end
-
     end
   end
 end